refactor(cli): refactor cli hook into cli module
I'm not currently convinced a separate hooks module is needed. I'm currently planning on definining hooks in their relevant modules, and then listing them all together in the documentation.
